feat: implementado hive para persistencia de dados locais, extraido tema para arquivo proprio
