+++ lib/configs/hive_config.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+import 'package:hive_flutter/hive_flutter.dart';
+import 'package:path_provider/path_provider.dart';
+
+class HiveConfig {
+  static Future<void> start() async {
+    Directory dir = await getApplicationDocumentsDirectory();
+    await Hive.initFlutter(dir.path);
+  }
+}
