+++ pubspec.yaml
@@ -36,6 +36,9 @@ dependencies:
   cupertino_icons: ^1.0.8
   intl: ^0.19.0
   provider: ^6.1.2
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+  path_provider: ^2.1.4
 
 dev_dependencies:
   flutter_test:
